refactor(pastoral-cycle): extract countdown widget to separate file
